(#289) Move parse_vars_conf to something_parser.hpp
